refactor(dsg): renaming the
