--- packages/amplication-data-service-generator/src/server/resource/dto/create-nested/index.ts
@@ -1,4 +1,0 @@
-export * from "./create-nested";
-export * from "./create-create-nested-many-without-input";
-export * from "./create-update-many-without-input";
-export * from "./nested-mutation-options";
