--- packages/amplication-data-service-generator/src/server/resource/dto/create-nested/nested-mutation-options.ts
@@ -1,5 +1,0 @@
-export enum NestedMutationOptions {
-  "Create" = "create",
-  "Connect" = "connect",
-  "ConnectOrCreate" = "connectOrCreate",
-}
