+++ packages/amplication-data-service-generator/src/server/resource/dto/nested-input-dto/create-create-nested-many-without-input/index.ts
