+++ packages/amplication-data-service-generator/src/server/resource/dto/nested-input-dto/create-update-many-without-input/create-update-many-without-input.ts
@@ -4,31 +4,31 @@ import { Entity } from "../../../../../types";
 import { NamedClassDeclaration } from "../../../../../util/ast";
 import { isToManyRelationField } from "../../../../../util/field";
 import { EntityDtoTypeEnum } from "../../entity-dto-type-enum";
-import { createNestedInputDTO } from "../create-nested";
+import { createNestedInputDTO } from "../create-nested-input-dto";
 
-export function createCreateNestedManyDTOs(
+export function createUpdateManyWithoutInputDTOs(
   entity: Entity
 ): NamedClassDeclaration[] {
   const toManyFields = entity.fields.filter(isToManyRelationField);
   return toManyFields.map((field) =>
     createNestedInputDTO(
-      createCreateNestedManyWithoutInputID(
+      createUpdateManyWithoutInputID(
         entity.pluralDisplayName,
         field.properties.relatedEntity.name
       ),
       entity,
       field,
-      EntityDtoTypeEnum.RelationCreateNestedManyWithoutSourceInput
+      EntityDtoTypeEnum.RelationUpdateManyWithoutSourceInput
     )
   );
 }
 
-export function createCreateNestedManyWithoutInputID(
+export function createUpdateManyWithoutInputID(
   pluralEntityName: string,
   nestedEntityName: string
 ): namedTypes.Identifier {
   return builders.identifier(
-    `${pascalCase(nestedEntityName)}CreateNestedManyWithout${pascalCase(
+    `${pascalCase(nestedEntityName)}UpdateManyWithout${pascalCase(
       pluralEntityName
     )}Input`
   );
