+++ packages/amplication-data-service-generator/src/server/resource/dto/nested-input-dto/create-update-many-without-input/index.ts
